pyproject.toml
[project]
name = "Comfly"
description = "NODES: Comfly_Mj, Comfly_mjstyle, Comfly_upload, Comfly_Mju, Comfly_Mjv, Comfly_kling_videoPreview"
version = "1.0.0"
license = {file = "LICENSE"}
dependencies = ["aiohttp", "aiohttp-cors", "GitPython", "numpy", "Pillow", "requests", "matrix-client", "transformers", "huggingface-hub", "psutil"]

[project.urls]
Repository = "https://github.com/ainewsto/Comfyui_Comfly"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "ainewsto"
DisplayName = "Comfyui_Comfly"
Icon = ""
